카메라 개념, 리지드바디가 없는 경우의 충돌 처리 진행중, 이 부분 꼼꼼히 봐야할듯

+++ D2DFramework/AnimationAsset.h
@@ -44,11 +44,11 @@ namespace d2dFramework
 	}
 	unsigned int AnimationAsset::GetAnimationSize() const
 	{
-		return mFrameInfomations.size();
+		return static_cast<unsigned int>(mFrameInfomations.size());
 	}
 	unsigned int AnimationAsset::GetFrameSize(unsigned int animationIndex) const
 	{
 		assert(animationIndex < GetAnimationSize());
-		return mFrameInfomations[animationIndex].size();
+		return static_cast<unsigned int>(mFrameInfomations[animationIndex].size());
 	}
 }
